# internal/server/api/openapi.yaml
openapi: "3.1.0"
info:
  title: express-botx API
  description: HTTP API for sending messages to eXpress corporate messenger via BotX API. Supports webhooks from Alertmanager and Grafana.
  version: dev
  license:
    name: MIT
    url: https://github.com/lavr/express-botx/blob/main/LICENSE
servers:
  - url: "{scheme}://{host}{basePath}"
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost:8080
      basePath:
        default: /api/v1
security:
  - bearerAuth: []
  - apiKeyHeader: []
paths:
...
  /command:
    post:
      operationId: command
      summary: BotX command callback
      description: |
        Receives command callbacks from the eXpress BotX platform.
        Parses the event type from the command body, routes to matching handler rules,
        and responds with 202 Accepted.

        This endpoint is registered only when `server.callbacks` is configured.
        Authentication uses JWT signed by the bot secret (Authorization: Bearer <jwt>),
        controlled by the `verify_jwt` setting.

        Event types are derived from the `command.body` field:
        - `system:chat_created` → `chat_created`
        - `system:added_to_chat` → `added_to_chat`
        - Plain text → `message`
        - etc.
      security:
        - callbackJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallbackPayload"
            example:
              sync_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              command:
                body: "system:chat_created"
                command_type: "system"
                data: {}
              from:
                user_huid: "f16cdc5f-6366-5552-9ecd-c36290ab3d11"
                group_chat_id: "30dc1980-643a-00ad-3c67-d5071b13e0d7"
                chat_type: "group_chat"
                host: "express.company.ru"
              bot_id: "054af49e-5e18-4dca-ad73-4f96b6de63fa"
              proto_version: 4
      responses:
        "202":
          description: Callback accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackAcceptedResponse"
        "400":
          description: Invalid request (malformed JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: JWT verification failed or missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <api-key>"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: "X-API-Key: <api-key>"
    botSignature:
      type: apiKey
      in: header
      name: X-Bot-Signature
      description: HMAC-SHA256 signature for bot secret auth
    callbackJWT:
      type: http
      scheme: bearer
      description: "JWT signed with bot secret (HS256). Used for callback endpoints from eXpress platform."
    gitlabToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
      description: "X-Gitlab-Token: <secret token> — secret token configured on the GitLab webhook. Matched against the default `server.gitlab.secret` or any `server.gitlab.senders[].secret` (a sender token scopes delivery to that sender's chats)."
  schemas:
...
    CallbackPayload:
      type: object
      description: BotX API v4 POST /command request body
      required: [sync_id, command, from, bot_id]
      properties:
        sync_id:
          type: string
          format: uuid
          description: Unique request identifier
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        command:
          $ref: "#/components/schemas/CallbackCommand"
        from:
          $ref: "#/components/schemas/CallbackFrom"
        bot_id:
          type: string
          format: uuid
          description: Bot identifier
          example: "054af49e-5e18-4dca-ad73-4f96b6de63fa"
        proto_version:
          type: integer
          description: Protocol version
          example: 4
        attachments:
          type: array
          description: Message attachments
          items:
            type: object
        async_files:
          type: array
          description: Metadata for asynchronously loaded files
          items:
            type: object
        entities:
          type: array
          description: Message entities (mentions, markup, etc.)
          items:
            type: object
    CallbackCommand:
      type: object
      description: Command part of a callback payload
      required: [body]
      properties:
        body:
          type: string
          description: |
            Command body text. System events use "system:" prefix
            (e.g. "system:chat_created"). Plain text means a regular message.
          example: "system:chat_created"
        command_type:
          type: string
          description: Command type identifier
          example: "system"
        data:
          type: object
          description: Additional command data
        metadata:
          type: object
          description: Command metadata
    CallbackFrom:
      type: object
      description: Sender information in a callback payload
      required: [group_chat_id]
      properties:
        user_huid:
          type: string
          format: uuid
          description: User HUID (may be empty for system events)
          example: "f16cdc5f-6366-5552-9ecd-c36290ab3d11"
        group_chat_id:
          type: string
          format: uuid
          description: Chat identifier
          example: "30dc1980-643a-00ad-3c67-d5071b13e0d7"
        chat_type:
          type: string
          description: Chat type
          example: "group_chat"
        host:
          type: string
          description: eXpress server host
          example: "express.company.ru"
        ad_chat:
          type: boolean
          description: Whether the chat is from Active Directory
        ad_group:
          type: boolean
          description: Whether the group is from Active Directory
...
    CallbackAcceptedResponse:
      type: object
      properties:
        result:
          type: string
          example: "accepted"
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "chat_id is required"